# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * GL Social API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  free: 'free',
  premium: 'premium',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  clerkId: string;
  username: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  country?: string | null;
  role: UserRole;
  coinBalance: number;
  isVerified: boolean;
  isBanned: boolean;
  ageVerified?: boolean;
  createdAt: string;
}

export interface UserUpdate {
  username?: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  country?: string | null;
}

export interface UserPage {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface AgeVerification {
  birthYear: number;
}

export interface Post {
  id: number;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  hashtags?: string[];
  authorId: number;
  author: User;
  likeCount: number;
  commentCount: number;
  isLiked: boolean;
  createdAt: string;
}

export interface PostInput {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface PostPage {
  posts: Post[];
  total: number;
  page: number;
  limit: number;
}

export interface LikeStatus {
  liked: boolean;
  likeCount: number;
}

export interface Comment {
  id: number;
  postId: number;
  content: string;
  authorId: number;
  author: User;
  createdAt: string;
}

export interface CommentInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface CoinBalance {
  userId: number;
  balance: number;
}

export type CoinTransactionType = typeof CoinTransactionType[keyof typeof CoinTransactionType];


export const CoinTransactionType = {
  credit: 'credit',
  debit: 'debit',
} as const;

export interface CoinTransaction {
  id: number;
  userId: number;
  amount: number;
  type: CoinTransactionType;
  description: string;
  createdAt: string;
}

export type RaffleStatus = typeof RaffleStatus[keyof typeof RaffleStatus];


export const RaffleStatus = {
  upcoming: 'upcoming',
  active: 'active',
  ended: 'ended',
} as const;

export interface Raffle {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  prize?: string | null;
  entryCost: number;
  startTime: string;
  endTime: string;
  status: RaffleStatus;
  entryCount: number;
  /** @nullable */
  winnerId?: number | null;
  winner?: User | null;
  hasEntered?: boolean;
  createdAt: string;
}

export interface RaffleInput {
  title: string;
  description: string;
  /** @nullable */
  prize?: string | null;
  /** @minimum 1 */
  entryCost: number;
  startTime: string;
  endTime: string;
}

export interface RaffleEntry {
  id: number;
  raffleId: number;
  userId: number;
  user: User;
  createdAt: string;
}

export interface RaffleWinner {
  raffleId: number;
  winner: User;
}

export type NotificationType = typeof NotificationType[keyof typeof NotificationType];


export const NotificationType = {
  raffle_started: 'raffle_started',
  raffle_ended: 'raffle_ended',
  raffle_won: 'raffle_won',
  post_liked: 'post_liked',
  post_commented: 'post_commented',
  coin_received: 'coin_received',
  system: 'system',
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  title: string;
  message: string;
  isRead: boolean;
  /** @nullable */
  referenceId?: number | null;
  createdAt: string;
}

export interface BanAction {
  banned: boolean;
  /** @nullable */
  reason?: string | null;
}

export interface CoinAdjustment {
  amount: number;
  description: string;
}

export interface PlatformStats {
  totalUsers: number;
  totalPosts: number;
  totalRaffles: number;
  activeRaffles: number;
  totalCoinsDistributed: number;
  newUsersToday?: number;
  newPostsToday?: number;
}

export interface FollowStatus {
  following: boolean;
  followerCount: number;
}

export interface SearchResults {
  users: User[];
  posts: Post[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListPostsParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
hashtag?: string | null;
/**
 * If true, only return posts from followed users
 */
following?: boolean;
};

export type ListCoinTransactionsParams = {
page?: number;
limit?: number;
};

export type ListRafflesParams = {
/**
 * @nullable
 */
status?: ListRafflesStatus;
};

export type ListRafflesStatus = typeof ListRafflesStatus[keyof typeof ListRafflesStatus] | null;


export const ListRafflesStatus = {
  active: 'active',
  upcoming: 'upcoming',
  ended: 'ended',
} as const;

export type SearchParams = {
q: string;
type?: SearchType;
};

export type SearchType = typeof SearchType[keyof typeof SearchType];


export const SearchType = {
  all: 'all',
  posts: 'posts',
  users: 'users',
} as const;

export type ListNotificationsParams = {
unreadOnly?: boolean;
};

export type AdminListUsersParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
search?: string | null;
};