# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GL Social API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get current user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /admin/users/{userId}/ban:
    patch:
      operationId: adminBanUser
      tags: [admin]
      summary: Ban or unban a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanAction"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/users/{userId}/coins:
    post:
      operationId: adminAddCoins
      tags: [admin]
      summary: Add GL Coins to a user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoinAdjustment"
      responses:
        "200":
          description: Updated balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinBalance"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, clerkId, username, role, coinBalance, isVerified, isBanned, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        username:
          type: string
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        role:
          type: string
          enum: [free, premium, admin]
        coinBalance:
          type: integer
        isVerified:
          type: boolean
        isBanned:
          type: boolean
        ageVerified:
          type: boolean
        createdAt:
          type: string
...
    UserPage:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    CoinBalance:
      type: object
      required: [userId, balance]
      properties:
        userId:
          type: integer
        balance:
          type: integer
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [raffle_started, raffle_ended, raffle_won, post_liked, post_commented, coin_received, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        referenceId:
          type: ["integer", "null"]
        createdAt:
          type: string
    BanAction:
      type: object
      required: [banned]
      properties:
        banned:
          type: boolean
        reason:
          type: ["string", "null"]
    CoinAdjustment:
      type: object
      required: [amount, description]
      properties:
        amount:
          type: integer
        description:
          type: string
...